% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cor_contigs_extract.R
\name{abundance_cor_cluster_filter}
\alias{abundance_cor_cluster_filter}
\title{Abundance Correlation Clustering Result Filtering}
\usage{
abundance_cor_cluster_filter(
  cor_threshold,
  sample_id,
  tpm_threshold,
  rdrp_multi_threshold = 100,
  non_rdrp_multi_threshold = 20
)
}
\arguments{
\item{cor_threshold}{Numeric value for correlation threshold.}

\item{sample_id}{Character string representing the sample ID.}

\item{tpm_threshold}{Numeric value for TPM threshold.}

\item{rdrp_multi_threshold}{Numeric value for RdRp multi threshold.}

\item{non_rdrp_multi_threshold}{Numeric value for non-RdRp multi threshold.}
}
\description{
This function filters the clustering results based on abundance correlation.
}
