% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient-matrix.R
\name{calculate_spearman_correlation}
\alias{calculate_spearman_correlation}
\title{Calculate the Spearman correlation between RdRp and other contigs}
\usage{
calculate_spearman_correlation(input_file, output_dir, threads = 20)
}
\arguments{
\item{input_file}{The input file prefix.}

\item{output_dir}{The directory to save the output files.}

\item{threads}{The number of CPU cores to use for parallel processing.}
}
\description{
This function processes the RdRp files, merges them, filters the data, and calculates the Spearman correlation.
}
