% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPM_combine.R
\name{process_rdrp_files}
\alias{process_rdrp_files}
\title{Process TPM files}
\usage{
process_rdrp_files(input_dir)
}
\arguments{
\item{input_dir}{The directory containing the RdRp files.}
}
\value{
A CSV file containing the merged TPM data.
}
\description{
This function merges the TPM of all contigs from all libraries and generates a CSV output.
}
