% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient-matrix_pre.R
\name{process_viruses_annotation}
\alias{process_viruses_annotation}
\title{Process Viruses Annotation}
\usage{
process_viruses_annotation(input_file, output_dir, length_threshold)
}
\arguments{
\item{input_file}{Path to the input file (without the \verb{_megahit_assemble_nr.edited.tsv} suffix).}

\item{output_dir}{Directory to save the output file.}

\item{length_threshold}{Length threshold for filtering.}
}
\description{
This function processes annotation information by replacing empty annotations with 'Viruses' and filtering based on length.
}
