% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastn_nt_novirus.R
\name{run_trim_contamination}
\alias{run_trim_contamination}
\title{Run Trim Contamination}
\usage{
run_trim_contamination()
}
\description{
This function parses command line arguments and runs the trim contamination process.
}
