% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastn_nt_novirus.R
\name{trim_contamination}
\alias{trim_contamination}
\title{Trim Contamination from Nucleotide Sequences}
\usage{
trim_contamination(input_file, db, out_tsv, threads, evalue, output_nt_fasta)
}
\arguments{
\item{input_file}{Path to the input nucleotide sequences file.}

\item{db}{Path to the BLAST database.}

\item{out_tsv}{Path to the output TSV file.}

\item{threads}{Number of threads to use.}

\item{evalue}{E-value threshold for BLAST.}

\item{output_nt_fasta}{Path to the output nucleotide FASTA file.}
}
\value{
Path to the output nucleotide FASTA file.
}
\description{
This function trims contamination from nucleotide sequences using BLAST.
}
